+++ src/main/java/com/wemakeitwork/allenvooreen/model/Team.java
@@ -4,14 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
+import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
@@ -27,7 +20,7 @@ public class Team {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @JsonProperty("id")
-    private long teamId;
+    private Integer teamId;
 
     @JsonProperty("name")
     private String teamName;
@@ -49,11 +42,11 @@ public class Team {
         this.medicationList.add(medication);
     }
 
-    public long getTeamId() {
+    public Integer getTeamId() {
         return teamId;
     }
 
-    public void setTeamId(long teamId) {
+    public void setTeamId(Integer teamId) {
         this.teamId = teamId;
     }
